refactor(organization): refactor actions & reducers
