+++ packages/web-app/src/components/appli/Form/OrganizationAutoComplete/CreateNewOrganization.jsx
@@ -7,7 +7,7 @@ import AddIcon from '@material-ui/icons/Add';
 import PropTypes from 'prop-types';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { postOrganization } from '../../../../actions/CreateOrganization';
+import { postOrganization } from '../../../../actions/Organization/CreateOrganization';
 import ActionButton from '../../../common/ActionButton';
 import { DocumentFormContext } from '../../Document/DocumentForm/Provider';
