+++ packages/web-app/src/reducers/CreateOrganization.js
@@ -2,7 +2,7 @@ import {
   POST_ORGANIZATION,
   POST_ORGANIZATION_FAILURE,
   POST_ORGANIZATION_SUCCESS
-} from '../actions/CreateOrganization';
+} from '../actions/Organization/CreateOrganization';
 
 const initialState = {
   error: null,
@@ -10,27 +10,26 @@ const initialState = {
   organization: null
 };
 
-const createOrganization = (state = initialState, action) => {
+const reducer = (state = initialState, action) => {
   switch (action.type) {
     case POST_ORGANIZATION:
       return {
         ...initialState,
         isLoading: true
       };
-    case POST_ORGANIZATION_SUCCESS:
-      return {
-        ...initialState,
-        isLoading: false,
-        organization: action.organization
-      };
     case POST_ORGANIZATION_FAILURE:
       return {
         ...initialState,
         error: action.error
+      };
+    case POST_ORGANIZATION_SUCCESS:
+      return {
+        ...initialState,
+        organization: action.organization
       };
     default:
       return state;
   }
 };
 
-export default createOrganization;
+export default reducer;
